requirements.txt
alembic==1.4.2
altair==4.1.0
amadeus==4.5.0
appdirs==1.4.4
appnope==0.1.0
argon2-cffi==20.1.0
asgiref==3.2.10
astor==0.8.1
astroid==2.4.2
attrs==20.1.0
backcall==0.2.0
base58==2.0.1
beautifulsoup4==4.9.1
bleach==3.1.5
blinker==1.4
bs4==0.0.1
cachetools==4.2.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
Django==3.1
entrypoints==0.3
fake-useragent==0.1.11
fastapi==0.61.0
filelock==3.0.12
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
gitdb==4.0.5
GitPython==3.1.11
h11==0.9.0
httptools==0.1.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
iniconfig==1.1.1
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==2.2.5
jupyterlab-server==1.2.0
lazy-object-proxy==1.4.3
lxml==4.5.2
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
moviepy==1.0.3
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.1.3
numpy==1.19.1
oauthlib==3.1.0
packaging==20.4
pandas==1.1.1
pandocfilters==1.4.2
parse==1.16.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pipenv==2020.8.13
plotly==4.14.1
pluggy==0.13.1
proglog==0.1.9
prometheus-client==0.8.0
prompt-toolkit==3.0.6
protobuf==3.14.0
psycopg2-binary==2.8.6
ptyprocess==0.6.0
py==1.9.0
pyarrow==2.0.0
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.6.1
pydeck==0.5.0
pyee==7.0.2
Pygments==2.6.1
pylint==2.5.3
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
pyppeteer==0.2.2
pyquery==1.4.1
pyrsistent==0.16.0
PySocks==1.7.1
pytest==6.1.2
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
pyzmq==19.0.2
qtconsole==4.7.6
QtPy==1.9.0
requests==2.24.0
requests-html==0.10.0
requests-oauthlib==1.3.0
retrying==1.3.3
selenium==3.141.0
Send2Trash==1.5.0
simplejson==3.17.2
six==1.15.0
smmap==3.0.4
soupsieve==2.0.1
SQLAlchemy==1.3.16
sqlparse==0.3.1
starlette==0.13.6
streamlit==0.72.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
toolz==0.11.1
tornado==6.0.4
tqdm==4.48.2
traitlets==4.3.3
tweepy==3.9.0
tzlocal==2.1
urllib3==1.25.10
uvicorn==0.11.8
uvloop==0.14.0
validators==0.18.1
virtualenv==20.0.30
virtualenv-clone==0.5.4
w3lib==1.22.0
watchdog==1.0.1
wcwidth==0.2.5
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
